Copy over token edit distance from metals
In the future I imagine we can turn this into a shared library. However,
I suspect I will need to customize this code to accommodate markdown
files because the positions inside code fences must be compared by their
offset in the original markdown document instead of positions inside the
code fence.
