+++ runtime/src/main/scala/vork/document/RangePosition.scala
@@ -1,6 +1,5 @@
 package vork.document
 
-
 object RangePosition {
   def empty: RangePosition = new RangePosition(-1, -1, -1, -1)
 }
