+++ vork/src/main/scala/vork/internal/markdown/MatchingToken.scala
@@ -1,0 +1,9 @@
+package vork.internal.markdown
+
+import scala.meta.Token
+
+/** A pair of tokens that align with each other across two different files */
+case class MatchingToken(original: Token, revised: Token) {
+  override def toString: String =
+    s"${original.structure} <-> ${revised.structure}"
+}
